+++ server.go
@@ -7,13 +7,16 @@ import (
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"encoding/pem"
+	"errors"
 	"log"
 	"math/big"
+	"os"
 	"time"
 
 	"github.com/luyuhuang/subsocks/server"
 	"github.com/luyuhuang/subsocks/utils"
 	"github.com/pelletier/go-toml"
+	"golang.org/x/crypto/ssh"
 )
 
 func launchServer(t *toml.Tree) {
@@ -66,10 +69,11 @@ func launchServer(t *toml.Tree) {
 	}
 
 	if config.Protocol == "ssh" {
-		ser.SSHConfig = &server.SSHConfig{
-			Cert: config.SSH.Cert,
-			Key:  config.SSH.Key,
+		sshConfig, err := getServerSSHConfig(config.SSH.Cert, config.SSH.Key, ser.Config)
+		if err != nil {
+			log.Fatalf("Get TLS configuration failed: %s", err)
 		}
+		ser.SSHConfig = sshConfig
 	}
 
 	if err := ser.Serve(); err != nil {
@@ -131,4 +135,56 @@ func genKeyPair() (rawCert, rawKey []byte, err error) {
 	rawKey = pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(priv)})
 
 	return
+}
+
+func getServerSSHConfig(cert string, key string, conf *server.Config) (*ssh.ServerConfig, error) {
+	// 公钥
+	authorizedKeysBytes, err := os.ReadFile(cert)
+	if err != nil {
+		return nil, err
+	}
+
+	authorizedKeysMap := map[string]bool{}
+	for len(authorizedKeysBytes) > 0 {
+		pubKey, _, _, rest, err := ssh.ParseAuthorizedKey(authorizedKeysBytes)
+		if err != nil {
+			return nil, err
+		}
+
+		authorizedKeysMap[string(pubKey.Marshal())] = true
+		authorizedKeysBytes = rest
+	}
+
+	// 服务端配置
+	config := &ssh.ServerConfig{
+		PasswordCallback: func(c ssh.ConnMetadata, pass []byte) (*ssh.Permissions, error) {
+			if conf.Verify(c.User(), string(pass)) {
+				return nil, nil
+			}
+			return nil, errors.New("ssh verify error: wrong username or password")
+		},
+		PublicKeyCallback: func(c ssh.ConnMetadata, pubKey ssh.PublicKey) (*ssh.Permissions, error) {
+			if authorizedKeysMap[string(pubKey.Marshal())] {
+				return &ssh.Permissions{
+					Extensions: map[string]string{
+						"pubkey-fp": ssh.FingerprintSHA256(pubKey),
+					},
+				}, nil
+			}
+			return nil, errors.New("ssh verify error: key authentication failed")
+		},
+	}
+
+	// 私钥
+	privateBytes, err := os.ReadFile(key)
+	if err != nil {
+		return nil, err
+	}
+	private, err := ssh.ParsePrivateKey(privateBytes)
+	if err != nil {
+		return nil, err
+	}
+	config.AddHostKey(private)
+
+	return config, nil
 }
