+++ server/ssh.go
@@ -3,7 +3,6 @@ package server
 import (
 	"errors"
 	"net"
-	"os"
 
 	"golang.org/x/crypto/ssh"
 )
@@ -64,56 +63,8 @@ func (s *sshStripper) Write(b []byte) (n int, err error) {
 }
 
 func (s *sshStripper) serverInit() (*ssh.Channel, error) {
-	// 公钥
-	authorizedKeysBytes, err := os.ReadFile(s.server.SSHConfig.Cert)
-	if err != nil {
-		return nil, err
-	}
-
-	authorizedKeysMap := map[string]bool{}
-	for len(authorizedKeysBytes) > 0 {
-		pubKey, _, _, rest, err := ssh.ParseAuthorizedKey(authorizedKeysBytes)
-		if err != nil {
-			return nil, err
-		}
-
-		authorizedKeysMap[string(pubKey.Marshal())] = true
-		authorizedKeysBytes = rest
-	}
-
-	// 服务端配置
-	config := &ssh.ServerConfig{
-		PasswordCallback: func(c ssh.ConnMetadata, pass []byte) (*ssh.Permissions, error) {
-			if s.server.Config.Verify(c.User(), string(pass)) {
-				return nil, nil
-			}
-			return nil, errors.New("ssh verify error: wrong username or password")
-		},
-		PublicKeyCallback: func(c ssh.ConnMetadata, pubKey ssh.PublicKey) (*ssh.Permissions, error) {
-			if authorizedKeysMap[string(pubKey.Marshal())] {
-				return &ssh.Permissions{
-					Extensions: map[string]string{
-						"pubkey-fp": ssh.FingerprintSHA256(pubKey),
-					},
-				}, nil
-			}
-			return nil, errors.New("ssh verify error: key authentication failed")
-		},
-	}
-
-	// 私钥
-	privateBytes, err := os.ReadFile(s.server.SSHConfig.Key)
-	if err != nil {
-		return nil, err
-	}
-	private, err := ssh.ParsePrivateKey(privateBytes)
-	if err != nil {
-		return nil, err
-	}
-	config.AddHostKey(private)
-
 	// 监听握手
-	_, chans, reqs, err := ssh.NewServerConn(s.Conn, config)
+	_, chans, reqs, err := ssh.NewServerConn(s.Conn, s.server.SSHConfig)
 	if err != nil {
 		return nil, err
 	}
